Split to base file.
